+++ Server/src/common/infrastructure/model/UserModel.cpp
@@ -1,6 +1,6 @@
 #include "UserModel.h"
 
-UserModel::UserModel(const unsigned int id, const std::string &username) :
+UserModel::UserModel(const long id, const std::string &username) :
     id(id),
     username(username)
 {}
